+++ node/src/main/scala/fluence/node/Configuration.scala
@@ -101,7 +101,7 @@ object Configuration extends slogging.LazyLogging {
     for {
       uid <- IO(scala.sys.process.Process("id -u").!!.trim)
       //TODO: don't do tendermint init if keys already exist
-      _ <- DockerIO.run[IO](tendermint("init", uid)).compile.drain
+      _ <- DockerIO.exec[IO](tendermint("init", uid))
 
       _ <- IO {
         tendermintDir.resolve("config").resolve("config.toml").toFile.delete()
@@ -109,10 +109,10 @@ object Configuration extends slogging.LazyLogging {
         tendermintDir.resolve("data").toFile.delete()
       }
 
-      nodeId <- DockerIO.run[IO](tendermint("show_node_id", uid)).compile.lastOrError
+      nodeId <- DockerIO.exec[IO](tendermint("show_node_id", uid))
       _ <- IO { logger.info(s"Node ID: $nodeId") }
 
-      validatorRaw <- DockerIO.run[IO](tendermint("show_validator", uid)).compile.lastOrError
+      validatorRaw <- DockerIO.exec[IO](tendermint("show_validator", uid))
       validator <- IO.fromEither(parse(validatorRaw).flatMap(_.as[ValidatorKey]))
       _ <- IO { logger.info(s"Validator PubKey: ${validator.value}") }
     } yield (nodeId, validator)
