+++ node/src/main/scala/fluence/node/MasterNodeApp.scala
@@ -16,19 +16,19 @@
 
 package fluence.node
 
+import cats.effect.ExitCase.{Canceled, Completed, Error}
 import cats.effect._
 import cats.syntax.functor._
 import com.softwaremill.sttp.SttpBackend
 import com.softwaremill.sttp.asynchttpclient.cats.AsyncHttpClientCatsBackend
 import fluence.ethclient.EthClient
+import fluence.node.config.SwarmConfig
 import fluence.node.eth.FluenceContract
 import fluence.node.workers.{CodeManager, SwarmCodeManager, TestCodeManager, WorkersPool}
 import fluence.swarm.SwarmClient
+import org.web3j.protocol.core.methods.response.EthSyncing.Syncing
 import slogging.MessageFormatter.DefaultPrefixFormatter
 import slogging.{LazyLogging, LogLevel, LoggerConfig, PrintLoggerFactory}
-import fluence.node.config.SwarmConfig
-import org.web3j.protocol.core.methods.response.EthSyncing.Syncing
-import cats.effect.ExitCase.{Canceled, Completed, Error}
 
 import scala.concurrent.duration._
 
@@ -96,7 +96,6 @@ object MasterNodeApp extends IOApp with LazyLogging {
             // Type annotations are here to make IDEA's type inference happy
             case (ethClient: EthClient, sttpBackend: SttpBackend[IO, Nothing], pool: WorkersPool[IO]) ⇒
               implicit val backend: SttpBackend[IO, Nothing] = sttpBackend
-              logger.info(s"Ethereum RPC config: $ethereumRpcConfig")
               for {
                 version ← ethClient.clientVersion[IO]()
                 _ = logger.info("eth client version {}", version)
@@ -120,10 +119,7 @@ object MasterNodeApp extends IOApp with LazyLogging {
           }
       }
       .attempt
-      .flatMap {
-        case Left(_) => IO.pure(ExitCode.Error)
-        case Right(ec) => IO.pure(ec)
-      }
+      .map(_.getOrElse(ExitCode.Error))
       .guaranteeCase {
         case Canceled =>
           IO(logger.error("MasterNodeApp was canceled"))
